Create CartService interface
