+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/entity/CartEntity.java
@@ -23,7 +23,7 @@ public class CartEntity {
     @JoinColumn(name = "user_id")
     private UserEntity userEntity;
 
-    @OneToMany(mappedBy = "cart", cascade = CascadeType.ALL, orphanRemoval = true)
+    @OneToMany(mappedBy = "cartEntity", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<CartItemEntity> items;
 
     @Column(name = "date_created", updatable = false)
